requirements.txt
pinecone-client==3.1.0
pinecone-text
FlagEmbedding
openai
flask
requests==2.26.0
gunicorn
numpy